requirements.txt
psycopg2-binary
pymongo
python-dotenv
matplotlib
pytz
bcrypt
requests
pytest